gprofng: updated man page and macro file for gp-display-gui
This is an update of the man page and the Texinfo macro definitions
file for the gprofng GUI.
The man page text has corrections for some option descriptions and
also ensures that the line lengths no longer extend beyond column 79.
The file with the Texinfo macros has one macro added.  It provides
a shorthand notation for the full name of the command to invoke the
GUI.

gprofng/ChangeLog
2023-12-01  Ruud van der Pas  <ruud.vanderpas@oracle.com>

	* doc/gp-display-gui.texi: Man page update for the gprofng GUI.
	* doc/gp-macros.texi: Added a new macro.
	* doc/local.mk: Set --release="gprofng-gui-$(VERSION)"